refactor(react-icons-font-subsetting-webpack-plugin): extract bundler types to own module
Moves the hand-maintained bundler interfaces out of `index.ts` into
`src/bundler-api.ts`, so the plugin module contains only subsetting logic and the
bundler surface is encapsulated in one place alongside the rationale for keeping
it hand-written.

`index.ts` re-exports the module (`export type * from './bundler-api'`), so the
package's public type surface is unchanged for consumers. `NormalModule` becomes
`BundlerNormalModule` now that it is exported, and the inline
`Compiler['webpack']['sources']['RawSource']` lookup becomes a named
`BundlerRawSource` alias.

`test/types.conformance.ts` now imports the interfaces from the new module, so the
drift check targets it directly.
